feat: update and successful build

+++ contracts/src/models/block.cairo
@@ -13,13 +13,13 @@ pub enum BlockType {
 // how can we get all all blocks in the area? → excute loop in the whole area? / set the index for
 // object?->hard for delete.
 #[derive(Copy, Drop, Serde)]
-#[dojo::model(namespace: "pixelaw", nomapping: true)]
+#[dojo::model]
 pub struct Block {
     #[key]
     pub stage_id: felt252,
     #[key]
-    pub x: u32,
+    pub x: u16,
     #[key]
-    pub y: u32,
+    pub y: u16,
     pub blocktype: BlockType
 }
